openrtb3: rename package to openrtb3 (avoid surprises)
