+++ openrtb3/body.go
@@ -1,4 +1,4 @@
-package openrtb
+package openrtb3
 
 // Body is a top-level wrapper for OpenRTB requests.
 type Body struct {
